+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Release workflow test failures - SettingsPage test now stubs `__APP_VERSION__` instead of hardcoding version (#710)
+
 ## [1.1.0] - 2026-01-11
 
 ### Added